requirements-dev.txt
django
djangorestframework
django-cors-headers
coverage
pylint==2.8.3
pre-commit